Extract tool failure handling
